ACP: extract dispatch lane and split manager domains
