+++ src/agents/acp-spawn.test.ts
@@ -172,9 +172,13 @@ describe("spawnAcpDirect", () => {
           agent: args.agent,
           runtimeSessionName,
           ...(cwd ? { runtimeOptions: { cwd }, cwd } : {}),
-          agentSessionId: "codex-inner-1",
-          backendSessionId: "acpx-1",
-          sessionIdsProvisional: true,
+          identity: {
+            state: "pending",
+            source: "ensure",
+            acpxSessionId: "acpx-1",
+            agentSessionId: "codex-inner-1",
+            lastUpdatedAt: Date.now(),
+          },
           mode: args.mode,
           state: "idle",
           lastActivityAt: Date.now(),
