+++ src/auto-reply/reply/commands-acp.test.ts
@@ -596,8 +596,13 @@ describe("/acp command", () => {
         backend: "acpx",
         agent: "codex",
         runtimeSessionName: "runtime-1",
-        backendSessionId: "acpx-sid-1",
-        agentSessionId: "codex-sid-1",
+        identity: {
+          state: "resolved",
+          source: "status",
+          acpxSessionId: "acpx-sid-1",
+          agentSessionId: "codex-sid-1",
+          lastUpdatedAt: Date.now(),
+        },
         mode: "persistent",
         state: "idle",
         lastActivityAt: Date.now(),
