+++ src/auto-reply/reply/commands-subagents-focus.test.ts
@@ -300,8 +300,13 @@ describe("/focus, /unfocus, /agents", () => {
         backend: "acpx",
         agent: "codex",
         runtimeSessionName: "runtime-1",
-        agentSessionId: "codex-123",
-        backendSessionId: "acpx-456",
+        identity: {
+          state: "resolved",
+          source: "status",
+          acpxSessionId: "acpx-456",
+          agentSessionId: "codex-123",
+          lastUpdatedAt: Date.now(),
+        },
         mode: "persistent",
         state: "idle",
         lastActivityAt: Date.now(),
