+++ src/config/sessions/types.ts
@@ -40,22 +40,6 @@ export type SessionAcpMeta = {
   agent: string;
   runtimeSessionName: string;
   identity?: SessionAcpIdentity;
-  /**
-   * @deprecated Compatibility projection. Use identity.acpxSessionId.
-   */
-  /** Backend-level ACP session identifier, if exposed by the runtime adapter. */
-  backendSessionId?: string;
-  /**
-   * @deprecated Compatibility projection. Use identity.agentSessionId.
-   */
-  /** Upstream harness session identifier (for example Codex/Claude session id), if exposed. */
-  agentSessionId?: string;
-  /**
-   * @deprecated Compatibility projection. Use identity.state.
-   * True while identifiers are still provisional (for example right after spawn, before first turn).
-   * When false/undefined, identifiers are treated as stable enough for user-facing surfaces.
-   */
-  sessionIdsProvisional?: boolean;
   mode: "persistent" | "oneshot";
   runtimeOptions?: AcpSessionRuntimeOptions;
   cwd?: string;
